+++ src/main/java/duke/ui/MainWindow.java
@@ -1,4 +1,4 @@
-package duke.gui;
+package duke.ui;
 
 import duke.DukeCore;
 import javafx.fxml.FXML;
