+++ src/main/java/duke/ui/PatientCard.java
@@ -1,4 +1,4 @@
-package duke.gui;
+package duke.ui;
 
 import duke.DukeCore;
 import duke.data.Patient;
