+++ src/main/java/duke/ui/PatientWindow.java
@@ -1,4 +1,4 @@
-package duke.gui;
+package duke.ui;
 
 import javafx.scene.layout.Region;
