+++ src/main/java/duke/ui/Ui.java
@@ -1,4 +1,4 @@
-package duke.gui;
+package duke.ui;
 
 import javafx.stage.Stage;
